refactor: edit spaces and use components in ui from desgin system
