+++ ui/src/main/java/com/baghdad/component/RecentSearchContent.kt
@@ -1,34 +1,36 @@
-package com.baghdad.components
+package com.baghdad.component
 
+import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
+import com.baghdad.design_system.component.Icon
+import com.baghdad.design_system.component.Text
 import com.baghdad.design_system.theme.Theme
 import com.baghdad.ui.R
 
 @Composable
 fun RecentSearchContent(
     title: String,
-    onCanceleClick: () -> Unit,
+    onCancelClick: () -> Unit,
     modifier: Modifier = Modifier
 ) {
     Row(
-        modifier = modifier,
+        modifier = modifier
+            .padding(vertical = 12.dp),
         verticalAlignment = Alignment.CenterVertically
     ) {
         Icon(
             painter = painterResource(id = com.baghdad.design_system.R.drawable.clock),
-            contentDescription = stringResource(R.string.clock_icon),
+            contentDescription = stringResource(id = R.string.clock_icon),
+            tint = Theme.color.hint,
             modifier = Modifier
                 .padding(end = 8.dp)
                 .size(20.dp)
@@ -39,17 +41,11 @@ fun RecentSearchContent(
             color = Theme.color.title
         )
         Spacer(modifier = Modifier.weight(1f))
-        IconButton(
-            onClick = { onCanceleClick() }) {
-            Icon(
-                painter = painterResource(id = com.baghdad.design_system.R.drawable.cancel),
-                contentDescription = stringResource(R.string.cancel_icon),
-                modifier = Modifier
-                    .padding(end = 12.dp)
-                    .size(16.dp)
-
-            )
-        }
+        Icon(
+            painter = painterResource(id = com.baghdad.design_system.R.drawable.cancel),
+            tint = Theme.color.hint,
+            contentDescription = stringResource(id = R.string.cancel_icon),
+            modifier = Modifier.clickable { onCancelClick() })
     }
 }
