+++ ui/src/main/java/com/baghdad/screen/SearchScreen.kt
@@ -3,7 +3,6 @@ package com.baghdad.screen
 import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.itemsIndexed
 import androidx.compose.runtime.Composable
@@ -11,11 +10,10 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.baghdad.components.HorizontalDivider
-import com.baghdad.components.MoviesCard
-import com.baghdad.components.RecentSearchContent
-import com.baghdad.components.SectionHeaderWithAction
-import com.baghdad.design_system.theme.NovixTheme
+import com.baghdad.component.HorizontalDivider
+import com.baghdad.component.MoviesCard
+import com.baghdad.component.RecentSearchContent
+import com.baghdad.component.SectionHeaderWithAction
 import com.baghdad.design_system.theme.Theme
 import com.baghdad.ui.R
 
@@ -32,7 +30,7 @@ fun SearchScreen(
     LazyColumn(
         modifier = Modifier
             .fillMaxSize()
-            .padding(horizontal = 16.dp)
+            .padding(vertical = 12.dp, horizontal = 16.dp)
     ) {
         item {
             SectionHeaderWithAction(
@@ -48,20 +46,19 @@ fun SearchScreen(
                 onSavedClick = { onSavedClick() },
                 modifier = Modifier
                     .padding(bottom = 12.dp)
-                    .width(168.dp)
+                    .fillParentMaxWidth(0.45f)
                     .aspectRatio(0.8f)
             )
         }
         item {
             SectionHeaderWithAction(
-                modifier = Modifier.padding(end = 12.dp),
                 title = stringResource(R.string.recent_search),
                 onClearAllClick = { onClearAllClick() })
         }
         itemsIndexed(recentSearches) { index, title ->
             RecentSearchContent(
                 title = title,
-                onCanceleClick = { onClearAllClick() }
+                onCancelClick = { onClearAllClick() }
             )
 
             if (index < recentSearches.lastIndex) {
@@ -80,7 +77,6 @@ fun SearchScreen(
 @Preview(showBackground = true, showSystemUi = true)
 @Composable
 private fun SearchScreenPreview() {
-    NovixTheme{
         SearchScreen(
             imageUrl = "",
             contentDescription = "",
@@ -89,5 +85,4 @@ private fun SearchScreenPreview() {
             onClearAllClick = {}
 
         )
-    }
 }
